pyproject.toml
[tool.poetry]
name = "crypto-divergence-scanner"
version = "0.1.0"
description = ""
authors = ["sanghun9634 <sanghun9634@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
pyzmq = "^27.0.0"
orjson = "^3.10.18"
pandas = "^2.3.0"
ta = "^0.11.0"
aiohttp = "^3.12.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
